Move Babel code to extension
